uptime-monitor/worker/requirements.txt
requests
psycopg2-binary
pytelegrambotapi
python-dotenv